better pg search for contributors and photo
