+++ django/apps/photo/models.py
@@ -7,7 +7,6 @@ import re
 from statistics import median
 
 from django.contrib.postgres.fields import JSONField
-from django.contrib.postgres.search import TrigramSimilarity
 from django.core.files.storage import default_storage
 from django.core.serializers.json import DjangoJSONEncoder
 from django.core.validators import FileExtensionValidator
@@ -16,13 +15,14 @@ from django.db.models.expressions import RawSQL
 from django.dispatch import receiver
 from django.utils import timezone
 from django.utils.translation import ugettext_lazy as _
-from model_utils.models import TimeStampedModel
-from slugify import Slugify
-from sorl.thumbnail import ImageField
 
 # from apps.issues.models import current_issue
 from apps.contributors.models import Contributor
 from apps.photo import file_operations
+from model_utils.models import TimeStampedModel
+from slugify import Slugify
+from sorl.thumbnail import ImageField
+from utils.dbfuncs import TrigramWordSimilarity
 from utils.merge_model_objects import merge_instances
 from utils.model_mixins import EditURLMixin
 
@@ -157,7 +157,7 @@ class ImageFileManager(models.Manager):
             return qs.filter(pk__in=pks)
 
         if filename:
-            trigram = TrigramSimilarity('stem', Path(filename).stem)
+            trigram = TrigramWordSimilarity('stem', Path(filename).stem)
             return qs.annotate(
                 similarity=trigram,
             ).filter(
